requirements.txt
flask
opencv-python-headless
pillow
rembg
numpy
python-dotenv
onnxruntime